hackumass/backend/requirements.txt
fastapi
uvicorn[standard]
snowflake-connector-python
python-dotenv
pydantic
google-generativeai
